Unzip almost done, added CPU count
Lots of fun stuff, except the case-insensitivity issues. I think I'd
rather write a routine for case-sensitive filesystems that check for
files with capital characters, and then prompts the user to lower-case
all file names. Then when I unzip P99 files, we can just lower-case the
destination, and be done with it.

miniz is nice in that it is in the public domain, and can be treated
like a library.